LayoutTests/js/arrowfunction-syntax-endings-expected.txt

Tests for ES6 arrow function endings

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS afEOL(12) is 13
PASS x=>x+1 did not throw exception.
PASS x=>x+1
 did not throw exception.
PASS f() is 39
PASS successfullyParsed is true

TEST COMPLETE
